Refactor tabbed workspace presenters
